test: move idle Full GC test into gc-controller-cadence.test.ts
Same controller, same directory; keeps GarbageCollectionController coverage
in one file.
